% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_rankings.R
\name{byor_etr}
\alias{byor_etr}
\title{Normalize ETR rankings}
\usage{
byor_etr(rankings, ...)
}
\arguments{
\item{rankings}{An ETR-shaped data frame.}

\item{...}{Arguments passed to \link{normalize_rankings}.}
}
\value{
A normalized ranking data frame.
}
\description{
Normalize ETR rankings
}
