% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_rankings.R
\name{byor_fantasypros}
\alias{byor_fantasypros}
\title{Normalize FantasyPros rankings}
\usage{
byor_fantasypros(rankings, ...)
}
\arguments{
\item{rankings}{A FantasyPros-shaped data frame.}

\item{...}{Arguments passed to \link{normalize_rankings}.}
}
\value{
A normalized ranking data frame.
}
\description{
Normalize FantasyPros rankings
}
