% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{fffloorceiling-package}
\alias{fffloorceiling}
\alias{fffloorceiling-package}
\title{Package imports}
\description{
Normalize fantasy football rankings, generate player outcome ranges with ffsimulator, and pass experimental team signals to nflseedR.
}
\keyword{internal}
